// translatorApp/src/app/types.ts
export enum OpenAIModels {
  'gpt-3.5-turbo',
  'gpt-4-1106-preview',
  'gpt-4',
}

export interface OpenAIResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: Array<{
    index: number;
    message: {
      role: string;
      content: string;
    };
    finish_reason: string;
  }>;
  usage: {
    prompt_tokens: number;
    completion_tokens: number;
    total_tokens: number;
  };
}